+++ src/AndroidStudio/localnotif/src/main/java/com/juankpro/ane/localnotif/Application.java
@@ -1,0 +1,14 @@
+package com.juankpro.ane.localnotif;
+
+/**
+ * Created by juancarlospazmino on 3/2/18.
+ */
+
+public class Application extends android.app.Application {
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        registerActivityLifecycleCallbacks(new LifecycleCallbacks());
+    }
+}