+++ src/AndroidStudio/testapplication/src/main/java/com/adobe/fre/FREExtension.java
@@ -1,0 +1,14 @@
+//
+// Source code recreated from a .class file by IntelliJ IDEA
+// (powered by Fernflower decompiler)
+//
+
+package com.adobe.fre;
+
+public interface FREExtension {
+    void initialize();
+
+    FREContext createContext(String var1);
+
+    void dispose();
+}