+++ src/AndroidStudio/testapplication/src/main/java/com/adobe/fre/FREFunction.java
@@ -1,0 +1,10 @@
+//
+// Source code recreated from a .class file by IntelliJ IDEA
+// (powered by Fernflower decompiler)
+//
+
+package com.adobe.fre;
+
+public interface FREFunction {
+    FREObject call(FREContext var1, FREObject[] var2);
+}