+++ src/AndroidStudio/testapplication/src/main/java/com/adobe/fre/FREWrongThreadException.java
@@ -1,0 +1,8 @@
+package com.adobe.fre;
+
+/**
+ * Created by juancarlospazmino on 3/12/18.
+ */
+
+public class FREWrongThreadException extends Exception {
+}